requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.1.1
langchain-community==0.3.27
langchain-openai==0.3.28
streamlit==1.47.0